backend/chatbot/requirements.txt
flask
flask-cors
python-dotenv
sentence-transformers
torch
pinecone-client==3.2.2
pypdf

